+++ lib/src/tab.dart
@@ -1,5 +1,4 @@
 import '../tabpanel.dart';
-import 'tab_panel.dart';
 import 'package:flutter/material.dart' hide Tab;
 import 'package:mobx/mobx.dart';