+++ lib/src/tab_panel_theme.dart
@@ -2,7 +2,6 @@ import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 
 final _defaultTabPanelTheme = const TabPanelThemeData();